+++ src/main/java/lotto/model/domain/LottoUser.java
@@ -1,5 +1,7 @@
 package lotto.model.domain;
 
+import lotto.validation.Validation;
+
 public class LottoUser {
     private Lottos lottos;
     private int money;
@@ -20,26 +22,9 @@ public class LottoUser {
 
     public LottoUser(Lottos lottos, int money, LottosWinningStatus lottosWinningStatus) {
         this.lottos = lottos;
-        validationMoney(money);
+        Validation.validationMoney(money);
         this.money = money;
         this.lottosWinningStatus = lottosWinningStatus;
-    }
-
-    private void validationMoney(int money) {
-        checkMoneyOutOfRange(money);
-        checkMoneyDivisibility(money);
-    }
-
-    private void checkMoneyDivisibility(int money) {
-        if (money % 1000 != 0) {
-            throw new IllegalArgumentException();
-        }
-    }
-
-    private void checkMoneyOutOfRange(int money) {
-        if (money < 0 || money > 100000) {
-            throw new IllegalArgumentException();
-        }
     }
 
     public Lottos getLottos() {
